Move processing to separate file
